Catch MalformedCSVError and treat as a validation error; RDR-407. (#307)
The current version of the CSV library appears to throw a CSV::MalformedCSVError when encountering an "invalid byte sequence in UTF-8" error parsing the manifest file
rather than an ArgumentError as was observed when the issue was originally noted.  If there turn out to be situations where an ArgumentError is thrown instead, then
we can deal with those at that time.

Also added a spec fixture CSV manifest file that contains an invalid UTF-8 character.  Didn't actually end up using it in a test but left it in the commit since it could
be useful for testing at some point.